refactor: reorganize handler/section package and folder structure
