+++ internal/repository/section/section_find_byId_test.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/stretchr/testify/require"
-	mocks "github.com/varobledo_meli/W17-G10-Bootcamp.git/mocks/section"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api/apperrors"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/section"
+	"github.com/varobledo_meli/W17-G10-Bootcamp.git/testhelpers"
 	"testing"
 )
 
@@ -28,7 +28,7 @@ func TestSectionRepository_FindById(t *testing.T) {
 		arrange
 		output
 	}
-	expSec := mocks.DummySection(1)
+	expSec := testhelpers.DummySection(1)
 	testCases := []testCase{
 		{
 			name: "returns sections by id",
