+++ internal/service/section/section_update_test.go
@@ -7,6 +7,7 @@ import (
 	service "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/service/section"
 	mocks "github.com/varobledo_meli/W17-G10-Bootcamp.git/mocks/section"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/section"
+	"github.com/varobledo_meli/W17-G10-Bootcamp.git/testhelpers"
 	"testing"
 )
 
@@ -29,8 +30,8 @@ func TestSectionDefault_UpdateSection(t *testing.T) {
 		output
 		input
 	}
-	sec := mocks.DummySection(1)
-	dumSec := mocks.DummySectionPatch(sec)
+	sec := testhelpers.DummySection(1)
+	dumSec := testhelpers.DummySectionPatch(sec)
 
 	testCases := []testCase{
 		{
@@ -39,11 +40,11 @@ func TestSectionDefault_UpdateSection(t *testing.T) {
 				func() *mocks.SectionRepositoryMock {
 					return &mocks.SectionRepositoryMock{
 						FuncFindById: func(ctx context.Context, id int) (*models.Section, error) {
-							dummySec := mocks.DummySection(1)
+							dummySec := testhelpers.DummySection(1)
 							return &dummySec, nil
 						},
 						FuncUpdate: func(ctx context.Context, id int, sec *models.Section) (*models.Section, error) {
-							dummySec := mocks.DummySection(1)
+							dummySec := testhelpers.DummySection(1)
 							return &dummySec, nil
 						},
 					}
